Fix Dex & Adapter Config
